+++ src/_types/data_row_T.d.ts
@@ -1,5 +1,5 @@
 import type { data_row_o_T } from './data_row_o_T'
-import type { data_row_tuple_T } from './data_row_tuple_T'
+import type { data_row__sym_props_o_T } from './data_row__sym_props_o_T'
 export type data_row_T<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
 > =
@@ -8,7 +8,7 @@ export type data_row_T<
 	: ColDefs extends [string, any][]
 		? data_row_tuple_T<ColDefs>&data_row_o_T<ColDefs>
 		: ColDefs extends any[]
-			? data_row_tuple_T<ColDefs>
+			? data_row_tuple_T<ColDefs>&data_row__sym_props_o_T<ColDefs>
 			: ColDefs extends object
 				? data_row_tuple_T<ColDefs>&data_row_o_T<ColDefs>
 				: never
@@ -18,14 +18,8 @@ export type Row<
 export type data_row_tuple_T<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
 > =
-	ColDefs extends (([string, any][])|any[]|object)
-	? string[]
-	: ColDefs extends [string, any][]
-		? ColDefs extends object
-			? string[]
-			: { [K in keyof ColDefs]:ColDefs[K][1] }
-		: ColDefs extends any[]
-			? ColDefs
-			: ColDefs extends object
-				? { [K in keyof ColDefs]:ColDefs[K][1] }
-				: never
+	ColDefs extends [string, any][]
+	? { [K in keyof ColDefs]:ColDefs[K][1] }
+	: ColDefs extends any[]
+		? ColDefs
+		: string[]