+++ src/data_row_/index.d.ts
@@ -1,10 +1,9 @@
-import type { data_row_T, data_row_tuple_T } from '../_types'
-import type { column_M_row_idx_T } from '../column_M_row_idx'
+import type { data_row_T, data_row_tuple_T, header_M_row_idx_T } from '../_types'
 export declare function data_row_<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
 >(
-	datum_a:data_row_tuple_T<ColDefs>,
-	column_M_column_idx:column_M_row_idx_T<ColDefs>
+	data_row_tuple:data_row_tuple_T<ColDefs>,
+	header_M_row_idx:header_M_row_idx_T<ColDefs>
 ):data_row_T<ColDefs>|undefined
 export {
 	data_row_ as _row_proxy,