+++ src/data_row_a/index.d.ts
@@ -1,6 +1,6 @@
 import type { ReadableAtom_ } from '@ctx-core/nanostores'
 import type { Ctx } from '@ctx-core/object'
-import type { data_row_T } from '../_types'
+import type { data_row_T, data_row_tuple_T, header_M_row_idx_T } from '../_types'
 export declare function data_row_a__<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
 >(ctx:Ctx):ReadableAtom_<data_row_T<ColDefs>[]>
@@ -12,6 +12,12 @@ export {
 export declare function data_row_a_<
 	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
 >(ctx:Ctx):data_row_T<ColDefs>[]
+export declare function data_row_a_<
+	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
+>(
+	data_row_tuple_a:data_row_tuple_T<ColDefs>[],
+	header_M_row_idx:header_M_row_idx_T<ColDefs>
+):data_row_T<ColDefs>[]
 export {
 	data_row_a_ as row_a_,
 }
@@ -21,3 +27,9 @@ export declare function data_row_a__set<
 export {
 	data_row_a__set as row_a__set,
 }
+export declare function data_row_a__new<
+	ColDefs extends (([string, any][])|any[]|object) = ([string, any][])|any[]|object
+>(
+	data_row_tuple_a:data_row_tuple_T<ColDefs>[],
+	header_M_row_idx:header_M_row_idx_T<ColDefs>
+):data_row_T<ColDefs>[]